tests/integration/requirements.txt
pytest>=7.0.0
pytest-asyncio>=0.21.0
google-cloud-firestore>=2.11.0
firebase-admin>=6.0.0